+++ path_tracing/src/math/mod.rs
@@ -1,4 +1,4 @@
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 use rand::Rng;
 use rand_distr::Uniform;
 
@@ -6,15 +6,15 @@ pub mod vector;
 
 pub struct Math;
 impl Math {
-    pub fn radian(degree: f32) -> f32 {
+    pub fn radian(degree: f64) -> f64 {
         return degree * PI / 180.0
     }
 
-    pub fn degree(radian: f32) -> f32 {
+    pub fn degree(radian: f64) -> f64 {
         return radian / PI * 180.0
     }
 
-    pub fn sample_uniform_distribution(low: f32, high: f32) -> f32 {
+    pub fn sample_uniform_distribution(low: f64, high: f64) -> f64 {
         let uni = Uniform::new(low, high);
         let mut rng = rand::thread_rng();
         return rng.sample(uni);
